fix: move v0.AttachedObjectReference helper functions back to client/v0/workload.go
